Separate survey task test files
